Clean up project structure and implement direct producer-consumer architecture with RabbitMQ
