+++ stock_service/stocks/apps.py
@@ -1,4 +1,5 @@
 import logging
+import sys
 from django.apps import AppConfig
 
 logger = logging.getLogger(__name__)
@@ -8,6 +9,17 @@ class StocksConfig(AppConfig):
     
     def ready(self):
         """
-        Initialize the stocks app when Django starts
+        Initialize the stocks app and start the RabbitMQ consumer when Django starts
         """
+        # Don't start the RabbitMQ consumer when running management commands
+        if 'runserver' in sys.argv:
+            # Import here to avoid AppRegistryNotReady exception
+            from .rabbitmq_consumer import start_rabbitmq_consumer
+            
+            logger.info("Starting RabbitMQ consumer for stock service...")
+            if start_rabbitmq_consumer():
+                logger.info("RabbitMQ consumer started successfully")
+            else:
+                logger.error("Failed to start RabbitMQ consumer")
+        
         logger.info("Stock service initialized")
